requirements/blog.txt
asgiref==3.3.4
crispy-forms-gds==0.2.2
Django==3.2
django-crispy-forms==1.11.2
django-datatables-too==0.1.3
django-environ==0.4.5
django-extra-views==0.13.0
django-multi-form-view==2.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
